feat: save allow/block list lookups
Previously list lookups would be done multiple times and also
for the special sentinel value address(0), which is not allowed
as a balance owner in the OZ ERC20 implementation.

This change means blocked and not-allowed addresses can
assign allowances and use permit signatures, however they
will not be able to receive or send funds.