feat: optimise oracle hot path
The anticipated hot functions are conversions, hence we want to minimize reads here.
We ue the sentinel value 0 (this is below the MIN_PRICE) while paused to fail
any conversions. To allow the admin to recover from any attack that might have manipulated
the prices, we need to keep shadow properties for prices that can be updated while paused
and moved back when unpaused.